fix: map for functions_ and types_
